mqms/requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
psycopg2-binary>=2.9
python-dotenv>=1.0
pandas>=2.0
openpyxl>=3.1
xhtml2pdf>=0.2.8
gunicorn>=20.1.0
django-cors-headers>=3.15.0
WeasyPrint>=58.2   # optional for nicer PDFs if you install system deps
